:tractor: Big refactor of internal package layout
