+++ cmd/parka/cmds/serve.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"github.com/go-go-golems/glazed/pkg/cli"
 	"github.com/go-go-golems/parka/pkg"
+	"github.com/go-go-golems/parka/pkg/glazed"
+	"github.com/go-go-golems/parka/pkg/render"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 	"io"
@@ -33,17 +35,17 @@ var ServeCmd = &cobra.Command{
 				Msg("Using assets from disk")
 			serverOptions = append(serverOptions,
 				pkg.WithStaticPaths(pkg.NewStaticPath(http.FS(os.DirFS("pkg/web/dist")), "/dist")),
-				pkg.WithPrependTemplateLookups(pkg.LookupTemplateFromDirectory("pkg/web/src/templates")),
+				pkg.WithPrependTemplateLookups(render.LookupTemplateFromDirectory("pkg/web/src/templates")),
 			)
 			cobra.CheckErr(err)
 		}
 
 		if templateDir != "" {
 			if dev {
-				serverOptions = append(serverOptions, pkg.WithPrependTemplateLookups(pkg.LookupTemplateFromDirectory(templateDir)))
+				serverOptions = append(serverOptions, pkg.WithPrependTemplateLookups(render.LookupTemplateFromDirectory(templateDir)))
 
 			} else {
-				lookup, err := pkg.LookupTemplateFromFS(os.DirFS(templateDir), ".", "**/*.tmpl.*")
+				lookup, err := render.LookupTemplateFromFS(os.DirFS(templateDir), ".", "**/*.tmpl.*")
 				cobra.CheckErr(err)
 				serverOptions = append(serverOptions, pkg.WithPrependTemplateLookups(lookup))
 			}
@@ -53,7 +55,7 @@ var ServeCmd = &cobra.Command{
 
 		s.Router.GET("/api/example", s.HandleSimpleQueryCommand(
 			NewExampleCommand(),
-			[]pkg.ParserHandlerOption{}))
+			[]glazed.ParserOption{}))
 		s.Router.POST("/api/example", s.HandleSimpleFormCommand(NewExampleCommand()))
 
 		err = s.Run()
